Fix issue with config deletion in utils test
